Implements block rm last cmd to delete the last block synced
This is done in order to easily fix the sync restart bug looking for the last "checkpoint". The sync process corrupts trying to restart and store a corrupted block. Removing last enables to try the process many times

+++ cmd/bitgodine/block/rm.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+	"github.com/xn3cr0nx/bitgodine_code/internal/blocks"
 	"github.com/xn3cr0nx/bitgodine_code/pkg/logger"
 )
 
@@ -16,9 +17,20 @@ var rmCmd = &cobra.Command{
 	Short: "Remove stored blocks",
 	Long:  "",
 	Run: func(cmd *cobra.Command, args []string) {
+		// height, err := strconv.Atoi(args[0])
+		// if err == nil {
+		// }
+		if args[0] == "last" {
+			if err := blocks.RemoveLast(); err != nil {
+				logger.Error("Block rm", err, logger.Params{})
+				os.Exit(-1)
+			}
+			os.Exit(1)
+		}
+
 		dir, err := ioutil.ReadDir(viper.GetString("dbDir"))
 		if err != nil {
-			logger.Error("transactions rm", err, logger.Params{})
+			logger.Error("Block rm", err, logger.Params{})
 		}
 		for _, d := range dir {
 			os.RemoveAll(path.Join([]string{viper.GetString("dbDir"), d.Name()}...))
